added new designation and client_type table and also added some new fileds in employee table

+++ EMS/Repository/ProjectRepo.cs
@@ -97,7 +97,7 @@ namespace EMS.Repository
                                 status = project.status,
                                 po = project.po,
                                 project_description = project.project_description,
-                                client_id = (int)project.client_id
+                                client_id = project.client_id
                             };
                 return query.ToList();
             }
